phase_3/pyproject.toml
[project]
name = "phase-3-planning"
version = "0.1.0"
description = "Phase 3: Planning agents"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.66.0",
    "python-dotenv>=1.2.2",
]

[tool.ruff]
line-length = 100
